Properly check of login_hook.login() exists and add regression tests
The login() function is not defined in the database extension sql any
more. This is because it will not be backed up if it is. So now it is
the user's responsibility to create a no-argument function called
login_hook.login(). If the login_hook schema exists (eg. the login_hook
extension is installed) but no function login_hook.login() exists, then
a message will be logged.