Added function login_hook.is_executing_login_hook()
The function login_hook.login() that is to be provided by the user must
have execution granted to public. It can invoke
login_hook.is_executing_login_hook() to check that it really is executed
as part of the login process, and not because someone thinks it nice to
invoke it randomly.