Made the installcheck more robust
The installcheck test didn't work on Postgres 11 because the proisagg
column doesn't exist any more in pg_catalog.pg_proc. This column has
nothing to do with the database extension but was used in the
installcheck test. Now the dependency has been removed to make the
installcheck pass on both Postgres 10 and Postgres 11.