#12: Segmentation fault with Postgres 14beta2
Apparently there is a problem using the SPI during the
session_preload_libraries phase. The login_hook.login() function will
now executed directly, so the SPI is taken out of the equation. This
helps: the function now executes as before.

But the login_hook.login() function may invoke other functions that do
use the SPI. This still is a problem.